+++ source/ProcessManager.Orchestrations.Tests/Integration/Processes/BRS_023_027/V1/MonitorOrchestrationUsingClientsScenario.cs
@@ -121,7 +121,8 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
                 CancellationToken.None);
 
         // Step 2: Wait service bus message to EDI and mock a response
-        await WaitAndMockServiceBusMessageToAndFromEdi(
+        await Fixture.EnqueueBrs023027ServiceBusListener.WaitAndMockServiceBusMessageToAndFromEdi(
+            ServiceProvider.GetRequiredService<IProcessManagerMessageClient>(),
             orchestrationInstanceId,
             calculationType);
 
@@ -214,7 +215,8 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
             .TriggerFunctionAsync("StartScheduledOrchestrationInstances");
 
         // Step 3: Wait service bus message to EDI and mock a response
-        await WaitAndMockServiceBusMessageToAndFromEdi(
+        await Fixture.EnqueueBrs023027ServiceBusListener.WaitAndMockServiceBusMessageToAndFromEdi(
+            ServiceProvider.GetRequiredService<IProcessManagerMessageClient>(),
             orchestrationInstanceId,
             calculationType);
 
@@ -289,41 +291,5 @@ public class MonitorOrchestrationUsingClientsScenario : IAsyncLifetime
             delay: TimeSpan.FromSeconds(3));
 
         isTerminated.Should().BeTrue("because we expects the orchestration instance can complete within given wait time");
-    }
-
-    private async Task WaitAndMockServiceBusMessageToAndFromEdi(
-        Guid orchestrationInstanceId,
-        CalculationType calculationType)
-    {
-        var verifyServiceBusMessage = await Fixture.EnqueueBrs023027ServiceBusListener
-            .When(
-                message =>
-                {
-                    if (message.Subject != $"Enqueue_{Brs_023_027.Name.ToLower()}")
-                        return false;
-
-                    var body = Energinet.DataHub.ProcessManager.Abstractions.Contracts.EnqueueActorMessagesV1
-                        .Parser.ParseJson(message.Body.ToString())!;
-
-                    var calculationCompleted = JsonSerializer.Deserialize<CalculatedDataForCalculationTypeV1>(body.Data);
-
-                    var typeMatches = calculationCompleted!.CalculationType == calculationType;
-                    var calculationIdMatches = calculationCompleted!.CalculationId == orchestrationInstanceId;
-                    var orchestrationIdMatches = body.OrchestrationInstanceId == orchestrationInstanceId.ToString();
-
-                    return typeMatches && calculationIdMatches && orchestrationIdMatches;
-                })
-            .VerifyCountAsync(1);
-        var messageFound = verifyServiceBusMessage.Wait(TimeSpan.FromSeconds(30));
-        messageFound.Should().BeTrue("because the expected message should be sent on the ServiceBus");
-
-        var processManagerMessageClient = ServiceProvider.GetRequiredService<IProcessManagerMessageClient>();
-
-        await processManagerMessageClient.NotifyOrchestrationInstanceAsync(
-            new NotifyOrchestrationInstanceEvent<NotifyEnqueueFinishedV1>(
-                OrchestrationInstanceId: orchestrationInstanceId.ToString(),
-                EventName: NotifyEnqueueFinishedV1.EventName,
-                Data: new NotifyEnqueueFinishedV1 { Success = true }),
-            CancellationToken.None);
     }
 }
